Made HandlerUtil class and its methods package private. Moved it to io.vertx.ext.sql package
