+++ src/main/java/io/vertx/ext/sql/SQLOperations.java
@@ -6,8 +6,6 @@ import io.vertx.codegen.annotations.VertxGen;
 import io.vertx.core.AsyncResult;
 import io.vertx.core.Handler;
 import io.vertx.core.json.JsonArray;
-import io.vertx.ext.sql.util.HandlerUtil;
-
 
 /**
  * Represents a SQL query interface to a database
